fix(pack): wasm runtime
