+++ .github/workflows/clippy.yml
@@ -19,7 +19,7 @@ jobs:
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           components: clippy
       # - name: Cache cargo
       #   uses: Swatinem/rust-cache@v2
