+++ .github/workflows/format.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           components: rustfmt
       - name: Check formatting
         run: cargo fmt --all -- --check
