+++ .github/workflows/pack-integration.yml
@@ -46,7 +46,7 @@ jobs:
       - name: Setup Rust
         uses: dtolnay/rust-toolchain@master
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           components: rustfmt, clippy
 
       # - name: Rust Cache
