+++ .github/workflows/pack-perf.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           components: rustfmt, clippy
 
       # - name: Cache cargo
