+++ .github/workflows/pack-release.yml
@@ -23,8 +23,8 @@ jobs:
             docker: ghcr.io/napi-rs/napi-rs/nodejs-rust:lts-debian-aarch64
             build: |
               apt-get update && apt-get install -y libc6-dev-arm64-cross libgcc-s1-arm64-cross &&
-              rustup install nightly-2026-02-05 &&
-              rustup default nightly-2026-02-05 &&
+              rustup install nightly-2026-02-18 &&
+              rustup default nightly-2026-02-18 &&
               rustup target add aarch64-unknown-linux-gnu &&
               npm run build:binding --workspace=@utoo/pack -- --target aarch64-unknown-linux-gnu
           - host: ubuntu-latest
@@ -38,8 +38,8 @@ jobs:
               echo "https://dl-cdn.alpinelinux.org/alpine/v3.21/community" >> /etc/apk/repositories &&
               apk update &&
               apk add --no-cache libc6-compat pkgconfig dav1d libdav1d dav1d-dev clang-static llvm-dev &&
-              rustup install nightly-2026-02-05 &&
-              rustup default nightly-2026-02-05 &&
+              rustup install nightly-2026-02-18 &&
+              rustup default nightly-2026-02-18 &&
               rustup target add aarch64-unknown-linux-musl &&
               export RUSTFLAGS='--cfg tokio_unstable -Zshare-generics=y -Zthreads=8 -Zunstable-options -Csymbol-mangling-version=v0 -Clinker-flavor=gnu-lld-cc -Clink-self-contained=+linker' &&
               npm run build:binding --workspace=@utoo/pack -- --target aarch64-unknown-linux-musl
@@ -48,8 +48,8 @@ jobs:
             docker: ghcr.io/napi-rs/napi-rs/nodejs-rust:lts-debian
             build: |
               apt-get update &&
-              rustup install nightly-2026-02-05 &&
-              rustup default nightly-2026-02-05 &&
+              rustup install nightly-2026-02-18 &&
+              rustup default nightly-2026-02-18 &&
               rustup target add x86_64-unknown-linux-gnu &&
               npm run build:binding --workspace=@utoo/pack -- --target x86_64-unknown-linux-gnu
           - host: ubuntu-latest
@@ -63,8 +63,8 @@ jobs:
               echo "https://dl-cdn.alpinelinux.org/alpine/v3.21/community" >> /etc/apk/repositories &&
               apk update &&
               apk add --no-cache libc6-compat pkgconfig dav1d libdav1d dav1d-dev clang-static llvm-dev &&
-              rustup install nightly-2026-02-05 &&
-              rustup default nightly-2026-02-05 &&
+              rustup install nightly-2026-02-18 &&
+              rustup default nightly-2026-02-18 &&
               rustup target add x86_64-unknown-linux-musl &&
               export RUSTFLAGS='--cfg tokio_unstable -Zshare-generics=y -Zthreads=8 -Csymbol-mangling-version=v0' &&
               npm run build:binding --workspace=@utoo/pack -- --target x86_64-unknown-linux-musl
@@ -109,7 +109,7 @@ jobs:
         uses: dtolnay/rust-toolchain@stable
         if: ${{ !matrix.settings.docker }}
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           targets: ${{ matrix.settings.target }}
       # - name: Cache cargo
       #   uses: Swatinem/rust-cache@v2
