+++ .github/workflows/pm-ci.yml
@@ -87,7 +87,7 @@ jobs:
         uses: dtolnay/rust-toolchain@stable
         if: ${{ !matrix.settings.docker }}
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           targets: ${{ matrix.settings.target }}
       
       # - name: Cache cargo
