+++ .github/workflows/pm-e2e.yml
@@ -65,7 +65,7 @@ jobs:
       - name: Setup Rust toolchain
         uses: dtolnay/rust-toolchain@stable
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           targets: ${{ matrix.settings.target }}
 
       # - name: Cache cargo
