+++ .github/workflows/pm-release.yml
@@ -94,7 +94,7 @@ jobs:
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           targets: ${{ matrix.target }}
 
       # - name: Cache cargo
