+++ .github/workflows/utooweb-ci.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
         with:
-          toolchain: nightly-2026-02-05
+          toolchain: nightly-2026-02-18
           targets: wasm32-unknown-unknown
           components: rust-src
       # - name: Cache cargo
