+++ crates/pack-api/Cargo.toml
@@ -10,8 +10,6 @@ anyhow              = { workspace = true, features = ["backtrace"] }
 bincode             = { workspace = true }
 either              = { workspace = true }
 futures             = { workspace = true }
-napi                = { workspace = true }
-napi-derive         = { workspace = true }
 parking_lot         = { workspace = true }
 qstring             = { workspace = true }
 rustc-hash          = { workspace = true }
@@ -47,7 +45,8 @@ vergen = { version = "9.0.6", features = ["cargo"] }
 [features]
 default      = ["process_pool"]
 process_pool = ["pack-core/process_pool", "turbopack-node/process_pool"]
-worker_pool  = ["pack-core/worker_pool", "turbopack-node/worker_pool_napi"]
+worker_pool      = ["pack-core/worker_pool", "turbopack-node/worker_pool"]
+worker_pool_napi = ["worker_pool", "pack-core/worker_pool_napi", "turbopack-node/worker_pool_napi"]
 
 [lints]
 workspace = true
