+++ crates/pack-napi/Cargo.toml
@@ -18,7 +18,7 @@ futures-util                 = { workspace = true }
 napi                         = { workspace = true }
 napi-derive                  = { workspace = true }
 owo-colors                   = { workspace = true }
-pack-api                     = { workspace = true }
+pack-api                     = { workspace = true, features = ["worker_pool_napi"] }
 pack-core                    = { workspace = true }
 rand                         = { workspace = true }
 rustc-hash                   = { workspace = true }
