+++ crates/pack-napi/src/pack_api/mod.rs
@@ -1,3 +1,4 @@
 pub mod endpoint;
 pub mod project;
+pub mod turbopack_ctx;
 pub mod utils;
